move apollo files
