+++ src/client/Apollo/queries.ts
@@ -1,0 +1,21 @@
+import { gql } from '@apollo/client';
+
+export const QUERY_ME = gql`
+	{
+		me {
+			_id
+			dateCreated
+			email
+			name {
+				name
+			}
+			picture
+		}
+	}
+`;
+
+export const HELLO_WORLD = gql`
+	{
+		helloWorld
+	}
+`;
