--- src/client/components/Apollo/queries.ts
@@ -1,15 +1,0 @@
-import gql from 'graphql-tag';
-
-export const QUERY_ME = gql`
-	{
-		me {
-			_id
-			dateCreated
-			email
-			name {
-        name
-      }
-			picture
-		}
-	}
-`;
